Sketch2Num/requirements.txt
altgraph==0.17.4
customtkinter==5.2.2
darkdetect==0.8.0
filelock==3.16.1
fsspec==2024.10.0
Jinja2==3.1.4
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
numpy==2.1.1
opencv-python==4.10.0.84
packaging==24.1
pefile==2024.8.26
pillow==10.4.0
pyinstaller==6.10.0
pyinstaller-hooks-contrib==2024.8
pywin32-ctypes==0.2.3
setuptools==74.1.2
sympy==1.13.1
tk==0.1.0
torch==2.5.1
torchvision==0.20.1
typing_extensions==4.12.2
